pyproject.toml
[project]
name = "snapclass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
